[Enhancement #338] Update test after moving method to ResourceDao
