+++ src/test/java/cz/cvut/kbss/termit/service/document/backup/ScheduledDocumentBackupManagerTest.java
@@ -3,9 +3,7 @@ package cz.cvut.kbss.termit.service.document.backup;
 import cz.cvut.kbss.jopa.model.EntityManager;
 import cz.cvut.kbss.termit.environment.Generator;
 import cz.cvut.kbss.termit.model.resource.File;
-import cz.cvut.kbss.termit.persistence.dao.ResourceDao;
 import cz.cvut.kbss.termit.service.document.BaseDocumentTestRunner;
-import cz.cvut.kbss.termit.util.Utils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -13,13 +11,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -33,9 +29,6 @@ public class ScheduledDocumentBackupManagerTest extends BaseDocumentTestRunner {
     private ScheduledDocumentBackupManager sut;
 
     @Autowired
-    private ResourceDao resourceDao;
-
-    @Autowired
     private EntityManager em;
 
     @MockitoBean
@@ -45,45 +38,18 @@ public class ScheduledDocumentBackupManagerTest extends BaseDocumentTestRunner {
 
     @BeforeEach
     public void generateFiles() {
-        Instant now = Utils.timestamp();
-
         modifiedAfterBackup = new ArrayList<>();
 
         transactional(() -> em.persist(document));
 
         // generate files with different modified timestamps
-        Map.of(now.minus(25, ChronoUnit.HOURS), now.minus(24, ChronoUnit.HOURS),
-                now.minus(23, ChronoUnit.HOURS), now.minus(25, ChronoUnit.HOURS),
-                now.minusSeconds(10), now.minus(1, ChronoUnit.MINUTES),
-                Instant.EPOCH, now.minus(5, ChronoUnit.DAYS))
-           .forEach((modified, lastBackup) -> {
-
-            File file = generateDocumentFile();
-            file.setModified(modified);
-            file.setLastBackup(lastBackup);
-            transactional(() -> em.persist(file));
-
-            if (modified.isAfter(lastBackup)) {
-                modifiedAfterBackup.add(file);
-            }
-        });
-    }
-
-    private File generateDocumentFile() {
-        java.io.File physicalFile = generateFile();
-        File file = new File();
-        file.setLabel(physicalFile.getName());
-        document.addFile(file);
-        file.setDocument(document);
-        file.setUri(Generator.generateUri());
-        return file;
-    }
-
-    @Test
-    void findFilesToBackupReturnsFilesModifiedAfterLastBackup() {
-        List<File> files = resourceDao.findModifiedFilesAfterLastBackup();
-        assertEquals(modifiedAfterBackup.size(), files.size());
-        assertTrue(files.containsAll(modifiedAfterBackup));
+        Generator.generateDocumentFilesWithBackupTimestamps(document)
+                .forEach(file -> {
+                    transactional(() -> em.persist(file));
+                    if (file.getModified().isAfter(file.getLastBackup())) {
+                        modifiedAfterBackup.add(file);
+                    }
+                });
     }
 
     @Test
@@ -93,6 +59,7 @@ public class ScheduledDocumentBackupManagerTest extends BaseDocumentTestRunner {
         ArgumentCaptor<File> fileCaptor = ArgumentCaptor.forClass(File.class);
         verify(backupManager, times(modifiedAfterBackup.size())).createBackup(fileCaptor.capture(), eq(BackupReason.SCHEDULED));
         List<File> files = fileCaptor.getAllValues();
+        assertFalse(files.isEmpty());
         assertEquals(modifiedAfterBackup.size(), files.size());
         assertTrue(files.containsAll(modifiedAfterBackup));
     }
